ml/requirements.txt
flask==2.0.1
requests
beautifulsoup4
yfinance==0.2.28
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
ta==0.7.0
vaderSentiment==3.3.2
scipy==1.10.1
groq==0.4.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.13
duckduckgo-search==3.9.3